Move appender lifecycle into appender_lifecycle.go
